+++ views/front/types/separator.php
@@ -18,7 +18,7 @@ $background = 'none';
 <?php if($active == 'yes'): ?>
     <?php $halfTextSize = $this->get('label_link') ? (strlen($this->get('label_link')) / 2) + 3 : 0 ?>
     <div class="separator" style="border-color: <?php echo $this->get('border_color', $page->get('separator_color')); ?>">
-        <div class="separator__line" style="max-width: calc(50% - <?php echo $halfTextSize ?>ch - .5px)">
+        <div class="separator__line" style="max-width: calc(50% - <?php echo $halfTextSize ?>ch - 1.5px)">
             <span style="background-color: <?php echo $this->get('border_color', $page->get('separator_color')); ?>"></span>
         </div>
         <?php if($this->get('label_link')): ?>
@@ -26,7 +26,7 @@ $background = 'none';
                 <?php echo $this->get('label_link'); ?>
             </div>
         <?php endif; ?>
-        <div class="separator__line" style="max-width: calc(50% - <?php echo $halfTextSize ?>ch - .5px)">
+        <div class="separator__line" style="max-width: calc(50% - <?php echo $halfTextSize ?>ch - 1.5px)">
             <span style="background-color: <?php echo $this->get('border_color', $page->get('separator_color')); ?>"></span>
         </div>
         <?php if($theme_id == 'full' || $theme_id == 'full-variant') { ?>