+++ views/links.php
@@ -15,6 +15,29 @@ $links              = WPLinkyHelper::getOptionValue('links', $data, []);
 $global             = WPLinkyHelper::getOptionValue('global', $data, []);
 ?>
 <div class="inside">
+    <div class="form-control mg-b-20">
+        <form
+                method="POST"
+                class="form-field _col-md-4 _col-xs-6 _js-links-form"
+                data-position="prepend"
+        >
+            <input type="hidden" name="_type" value="default">
+            <button type="submit" class="button button-new button-large"><?php echo __('Add link', 'linky'); ?></button>
+        </form>
+        <form
+                method="POST"
+                class="form-field _col-md-4 _col-xs-6 _js-links-form"
+                data-position="prepend"
+        >
+            <input type="hidden" name="_type" value="separator">
+            <button type="submit" class="button button-new button-large"><?php echo __('Add separator', 'linky'); ?></button>
+        </form>
+        <div class="form-field _col-md-4 _col-xs-12">
+            <button type="submit" class="button button-primary button-large" form="links"><?php echo __('Save'); ?></button>
+        </div>
+        <div class="clearfix"></div>
+    </div>
+
     <form
             method="POST"
             action="<?php echo admin_url( 'admin-ajax.php' ); ?>"